modified:   a11y-configs/generate-run-scripts.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/110-install-cower.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/120-build-espeak.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/130-build-speech-dispatcher.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/140-build-speechd-up.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/150-install-tclx.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/160-install-emacspeak.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/170-install-brltty.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/220-install-rootfs-expansion-scripts.sh modified:   a11y-configs/pi-arch/console-sd-libao/scripts/230-install-release-files.sh

+++ a11y-configs/pi-arch/console-sd-libao/scripts/110-install-cower.sh
@@ -8,11 +8,14 @@ tar -zxf cower-git.tar.gz
 pushd cower-git
 sed -i-old "s:^arch=(\(.*\)):arch=(\1 'armv6h'):" PKGBUILD
 makepkg -i --asroot --noconfirm --noprogressbar
-echo '-- Finished making and installing cower, tidying up...' 
 popd
-set +e
-rm -rf cower-git/
-rm cower-git.tar.gz
+echo '-- Finished making and installing cower, tidying up...' 
+if [ "${SM_TIDY}" ]; then
+	set +e
+	rm -rf cower-git/
+	rm cower-git.tar.gz
+fi
+
 exit 0
